+++ Text/Megaparsec/Prim.hs
@@ -31,6 +31,7 @@ module Text.Megaparsec.Prim
   , unexpected
   , (<?>)
   , label
+  , hidden
   , try
   , lookAhead
   , notFollowedBy
@@ -466,6 +467,12 @@ label p l = ParsecT $ \s cok cerr eok eerr ->
       eok' x s' hs = eok x s' $ refreshLastHint hs l
       eerr'    err = eerr $ setErrorMessage (Expected l) err
   in unParser p s cok' cerr eok' eerr'
+
+-- | @hidden p@ behaves just like parser @p@, but it doesn't show any “expected”
+-- tokens in error message when @p@ fails.
+
+hidden :: ParsecT s u m a -> ParsecT s u m a
+hidden p = label p ""
 
 -- | The parser @try p@ behaves like parser @p@, except that it
 -- pretends that it hasn't consumed any input when an error occurs.